AI-Service/pyproject.toml
[project]
name = "model-serving"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.11",
    "locust>=2.33.0",
    "numpy>=2.2.3",
    "pillow>=11.1.0",
    "ray[all]>=2.43.0",
    "requests>=2.32.3",
    "transformers>=4.49.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "accelerate>=0.26.0",
    "python-dotenv>=1.0.1",
    "sentencepiece>=0.2.0",
]


[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
